API for updating the reservation
Currently it simple checks the criteria for updating the reservation such as checking the overlapping. TODO in this API. **. Before adding the flight, check if the enough seats are avaible in the flight to be addedd 1. Update the passenger list in the added and removed flights 2. Update the price in resevation. 3. Update the seatsLeft in the added and removed flights